renamed page to view since view can be part of a page
page is an overloaded term
view is a container for pageobjects
when container is browser view can be entire page.
but container can be a panel on a page, some div.
When scope of container is less than browser it makes sense to call this a view and not a page.
Hence the reason.